fix: get java impl to pass the new sax encoding tests
